+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGE LOG
 
+## main branch
+
+- `telegram bot`  白名单配置
+- `ENABLE_WEBHOOK` 添加 webhook
+- admin 页面使用双层 tab
+
 ## v0.4.2
 
 - 修复 smtp imap proxy sever 的一些 bug